+++ blocks/map/block.ts
@@ -1,10 +1,66 @@
-export {};
+import { IconName } from '@fortawesome/fontawesome-common-types';
+
+interface AbstractMarkerIcon {
+  /**
+   * The anchor X and Y offset used for positioning the image.
+   *
+   * By default, the center of the icon will be used to mark the location. For many icons, it may
+   * be desirable to customize this. For example, for a symmetric pin which has a width of 10, and a
+   * height of 16, you’ll probably want to set this to `[5, 16]`
+   *
+   * The following special cases for Font Awesome icons are treated in a special way, since they are
+   * often used to represent a location:
+   *
+   * - `map-marker`
+   * - `map-marker-alt`
+   * - `map-pin`
+   * - `thumbtrack`
+   */
+  anchor?: [number, number];
+
+  /**
+   * The height of marker icons in pixels.
+   *
+   * @default 28
+   */
+  size?: number;
+
+  /**
+   * Enlarge an active marker with this ratio.
+   *
+   * This modifier is applied if a marker matches the block data which may be received by the block
+   * context. For example, this is applied if a marker matches the data that was passed in from a
+   * link action.
+   *
+   * @default 1
+   */
+  activeRatio?: number;
+}
+
+interface FontAwesomeMarkerIcon extends AbstractMarkerIcon {
+  /**
+   * A Font Awesome icon name to use.
+   */
+  icon?: IconName;
+}
+
+interface AssetMarkerIcon extends AbstractMarkerIcon {
+  /**
+   * The id of an asset to use.
+   */
+  asset: number;
+}
 
 declare module '@appsemble/sdk' {
   interface Parameters {
     latitude: string;
     longitude: string;
     disableClustering?: boolean;
+
+    /**
+     * Custom icon configuration.
+     */
+    icons?: FontAwesomeMarkerIcon | AssetMarkerIcon;
 
     /**
      * The maximum radius that a cluster will cover from the central marker (in pixels). Default 80.