+++ blocks/map/src/loadMarkers.ts
@@ -1,13 +1,8 @@
 import { BootstrapParams } from '@appsemble/sdk';
-import { Icon, LatLngBounds, LayerGroup, Map, Marker } from 'leaflet';
+import { LatLngBounds, LayerGroup, Map, Marker } from 'leaflet';
 
-import iconUrl from '../../../themes/amsterdam/core/marker.svg';
 import { LatLngMapper } from './createGetters';
-
-const MARKER_ICON_WIDTH = 39;
-const MARKER_ICON_HEIGHT = 39;
-const ACTIVE_MARKER_ICON_WIDTH = 64;
-const ACTIVE_MARKER_ICON_HEIGHT = 64;
+import createIcon from './createIcon';
 
 export function makeFilter(fields: [string, string], bounds: LatLngBounds): string {
   const [lon, lat] = fields;
@@ -28,13 +23,13 @@ export default function loadMarkers(
   fetched: Set<number>,
   get: LatLngMapper,
   data: any,
-  actions: BootstrapParams['actions'],
+  params: BootstrapParams,
   target: LayerGroup | Map,
 ): void {
   if (!Array.isArray(markers)) {
     return;
   }
-  markers.forEach((marker) => {
+  markers.forEach(async (marker) => {
     if (fetched.has(marker.id)) {
       return;
     }
@@ -44,21 +39,8 @@ export default function loadMarkers(
     if (Number.isNaN(Number(lat)) || Number.isNaN(Number(lng))) {
       return;
     }
-    const m = new Marker([lat, lng], {
-      icon:
-        data && data.id === marker.id
-          ? new Icon({
-              iconUrl,
-              iconSize: [ACTIVE_MARKER_ICON_WIDTH, ACTIVE_MARKER_ICON_HEIGHT],
-              iconAnchor: [ACTIVE_MARKER_ICON_WIDTH / 2, ACTIVE_MARKER_ICON_HEIGHT],
-            })
-          : new Icon({
-              iconUrl,
-              iconSize: [MARKER_ICON_WIDTH, MARKER_ICON_HEIGHT],
-              iconAnchor: [MARKER_ICON_WIDTH / 2, MARKER_ICON_HEIGHT],
-            }),
-    });
-    m.on('click', actions.onMarkerClick.dispatch.bind(null, marker));
-    target.addLayer(m);
+    new Marker([lat, lng], { icon: await createIcon(params, data && data.id === marker.id) })
+      .on('click', params.actions.onMarkerClick.dispatch.bind(null, marker))
+      .addTo(target);
   });
 }